Update klicky-macros.cfg
Address an issue with blindly assuming that if Z is already homed then it is above safe_z which was overwriting the value of safe_z to the current Z position when re-homing if Z was already homed regardless of whether the current position was actually above the safe_z. Additionally  included small typo fixes for logs.